requirements.txt
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.28.0

